All Pieces Thrown in
added all possible pieces. Checked generation. all is good

next expect save state of board - for saving the objects in the dict - Need to figure out a way to get enumerate to work properly. or get this length function to work. maybe read in the key as a string rather than 'key' varible. or an int.